Made data_frame checks column-based, renamed files.
